Split library classes
