+++ src/lib/Count64.php
@@ -142,155 +142,6 @@ function unpack64le($data) {
   ));
 }
 
-abstract class Count64Base {
-  protected $limit32Bit = False;
-
-  function __construct($value = 0, $limit32Bit = False) {
-    $this->limit32Bit = $limit32Bit;
-    $this->set($value);
-  }
-
-  abstract public function set($value);
-  abstract public function add($value);
-  abstract public function getHiBytes();
-  abstract public function getLoBytes();
-  abstract public function _getValue();
-
-  const EXCEPTION_SET_INVALID_ARGUMENT = "Count64 object can only be set() to integer or Count64 values";
-  const EXCEPTION_ADD_INVALID_ARGUMENT = "Count64 object can only be add()ed integer or Count64 values";
-  const EXCEPTION_32BIT_OVERFLOW = "Count64 object limited to 32 bit (overflow)";
-}
-
-class Count64_32 extends Count64Base {
-  private $loBytes;
-  private $hiBytes;
-
-  public function getHiBytes() {
-    return $this->hiBytes;
-  }
-
-  public function getLoBytes() {
-    return $this->loBytes;
-  }
-
-  public function _getValue() {
-    return array($this->hiBytes, $this->loBytes);
-  }
-
-  public function set($value) {
-    if (is_int($value)) {
-      $this->loBytes = $value;
-      $this->hiBytes = 0;
-    } else if (is_array($value) && 2 == sizeof($value)) {
-      $this->loBytes = $value[0];
-      if ($this->limit32Bit && 0 !== $value[1]) {
-        throw new \OverflowException(self::EXCEPTION_32BIT_OVERFLOW);
-      }
-      $this->hiBytes = $value[1];
-    } else if (is_object($value) && __CLASS__ == get_class($value)) {
-      $value = $value->_getValue();
-          if ($this->limit32Bit && 0 !== $value[0]) {
-        throw new \OverflowException(self::EXCEPTION_32BIT_OVERFLOW);
-      }
-      $this->hiBytes = $value[0];
-      $this->loBytes = $value[1];
-    } else {
-      throw new \InvalidArgumentException(self::EXCEPTION_SET_INVALID_ARGUMENT);
-    }
-    return $this;
-  }
-
-  public function add($value) {
-    if (is_int($value)) {
-      $sum = (int) ($this->loBytes + $value);
-      // overflow!
-      if (($this->loBytes > -1 && $sum < $this->loBytes && $sum > -1)
-        || ($this->loBytes < 0 && ($sum < $this->loBytes || $sum > -1))) {
-        if ($this->limit32Bit) {
-          throw new \OverflowException(self::EXCEPTION_32BIT_OVERFLOW);
-        }
-        $this->hiBytes = (int) ($this->hiBytes + 1);
-      }
-      $this->loBytes = $sum;
-    } else if (is_object($value) && __CLASS__ == get_class($value)) {
-      $value = $value->_getValue();
-      $sum = (int) ($this->loBytes + $value[1]);
-      if (($this->loBytes > -1 && $sum < $this->loBytes && $sum > -1)
-        || ($this->loBytes < 0 && ($sum < $this->loBytes || $sum > -1))) {
-        if ($this->limit32Bit) {
-          throw new \OverflowException(self::EXCEPTION_32BIT_OVERFLOW);
-        }
-        $this->hiBytes = (int) ($this->hiBytes + 1);
-      }
-      $this->loBytes = $sum;
-      if ($this->limit32Bit && 0 !== $value[0]) {
-        throw new \OverflowException(self::EXCEPTION_32BIT_OVERFLOW);
-      }
-      $this->hiBytes = (int) ($this->hiBytes + $value[0]);
-    } else {
-      throw new \InvalidArgumentException(self::EXCEPTION_ADD_INVALID_ARGUMENT);
-    }
-    return $this;
-  }
-}
-
-class Count64_64 extends Count64Base {
-  private $value;
-
-  public function getHiBytes() {
-    return urShift($this->value, 32);
-  }
-
-  public function getLoBytes() {
-    return $this->value & INT64_LOW_MAP;
-  }
-
-  public function _getValue() {
-    return $this->value;
-  }
-
-  public function set($value) {
-    if (is_int($value)) {
-      if ($this->limit32Bit && INT_MAX_32 < $value) {
-        throw new \OverFlowException(self::EXCEPTION_32BIT_OVERFLOW);
-      }
-      $this->value = $value;
-    } else if (is_array($value) && 2 == sizeof($value)) {
-      if ($this->limit32Bit && 0 !== $value[1]) {
-        throw new \OverFlowException(self::EXCEPTION_32BIT_OVERFLOW);
-      }
-      $this->value = $value[1];
-      $this->value = $this->value << 32;
-      $this->value = $this->value + $value[0];
-    } else if (is_object($value) && __CLASS__ == get_class($value)) {
-      $value = $value->_getValue();
-      if ($this->limit32Bit && INT_MAX_32 < $value) {
-        throw new \OverFlowException(self::EXCEPTION_32BIT_OVERFLOW);
-      }
-      $this->value = $value;
-
-    } else {
-      throw new \InvalidArgumentException(self::EXCEPTION_SET_INVALID_ARGUMENT);
-    }
-    return $this;
-  }
-
-  public function add($value) {
-    if (is_int($value)) {
-      $sum = (int) ($this->value + $value);
-    } else if (is_object($value) && __CLASS__ == get_class($value)) {
-      $sum = (int) ($this->value + $value->_getValue());
-    } else {
-      throw new \InvalidArgumentException(self::EXCEPTION_ADD_INVALID_ARGUMENT);
-    }
-    if ($this->limit32Bit && INT_MAX_32 < $sum) {
-      throw new \OverFlowException(self::EXCEPTION_32BIT_OVERFLOW);
-    }
-    $this->value = $sum;
-    return $this;
-  }
-}
-
 abstract class Count64 {
   public static function construct($value = 0, $limit32Bit = False) {
     if (4 == PHP_INT_SIZE) {
